Extract availability checks
Pull the availability check logic out into its own class to clean up the
`SignUp` model.
